wip cli test
