+++ CMakeLists.txt
@@ -65,7 +65,7 @@ target_compile_options(commandershared PRIVATE ${flags})
 target_compile_options(commandlineshared PRIVATE ${flags})
 
 # defines executables to build
-add_executable(command-line-test ./src/utils/command_line_test.c)
+# add_executable(cli-test ./src/cli_test.c)
 add_executable(c-template-cli-REPLACEME ./src/cli.c)
 add_executable(logger-test ./src/utils/logger.c)
 add_executable(array-len-test ./src/utils/array_test.c)
@@ -74,7 +74,7 @@ add_executable(safe-math-test ./src/utils/safe_math_test.c)
 add_executable(safe-mem-test ./src/utils/safe_mem_test.c)
 
 # set executable compiler options
-target_compile_options(command-line-test PRIVATE ${flags})
+# target_compile_options(cli-test PRIVATE ${flags})
 target_compile_options(c-template-cli-REPLACEME PRIVATE ${flags})
 target_compile_options(logger-test PRIVATE ${flags})
 target_compile_options(array-len-test PRIVATE ${flags})
@@ -83,12 +83,12 @@ target_compile_options(safe-math-test PRIVATE ${flags})
 target_compile_options(safe-mem-test PRIVATE ${flags})
 
 # links libraries
-target_link_libraries(command-line-test cmocka)
-target_link_libraries(command-line-test commandlineshared)
-target_link_libraries(command-line-test safememstatic)
-target_link_libraries(command-line-test loggershared)
-target_link_libraries(command-line-test zlog)
-target_link_libraries(command-line-test commandershared)
+# target_link_libraries(cli-test cmocka)
+# target_link_libraries(cli-test commandlineshared)
+# target_link_libraries(cli-test safememstatic)
+# target_link_libraries(cli-test loggershared)
+# target_link_libraries(cli-test zlog)
+# target_link_libraries(cli-test commandershared)
 target_link_libraries(c-template-cli-REPLACEME commandlineshared)
 target_link_libraries(c-template-cli-REPLACEME safememstatic)
 target_link_libraries(c-template-cli-REPLACEME loggershared)
@@ -107,7 +107,7 @@ target_link_libraries(safe-mem-test safememstatic)
 
 # enable testing
 enable_testing()
-add_test(NAME CommandLineTest COMMAND command-line-test)
+add_test(NAME CommandLineTest COMMAND cli-test)
 add_test(NAME ArrayLenTest COMMAND array-len-test)
 add_test(NAME RandomTest COMMAND random-test)
 add_test(NAME SafeMathTest COMMAND safe-math-test)
